#122: Smoothed borders

+++ Source/Gorgon/Graphics/BitmapFont.h
@@ -229,7 +229,7 @@ namespace Gorgon { namespace Graphics {
 
 		virtual int GetDigitWidth() const override { return digw; }
 
-		virtual int GetLineThickness() const override { return linethickness; }
+		virtual float GetLineThickness() const override { return linethickness; }
 
 		virtual int GetUnderlineOffset() const override { return underlinepos; }
